25/10/2023 20:06

+++ Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/entity/WeeklyEmptySlot.java
@@ -20,7 +20,7 @@ public class WeeklyEmptySlot {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "weekly_slot_id")
-    private int weeklySlotId;
+    private Long weeklySlotId;
 
     @Column(name = "first_day_of_week", nullable = false)
     private Date firstDayOfWeek;
